third_party/blink/web_tests/external/wpt/wasm/webapi/esm-integration/wasm-import.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Invalid imports for WebAssembly modules should error.
  assert_equals: expected (number) 1 but got (string) "unexpected"
Harness: the test ran to completion.
